simulate_producer_consumer/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
asyncio==3.4.3 